requirements.txt
Django==1.10.5
django-webpack-loader==0.4.1
djangorestframework==3.5.4
django-rest-auth==0.9.0
